ports/PWM: Change the PWM examples in the documentation.
Adding the freq and duty_u16 keyword settings to the constructor and
sometimes other details in the PWM section.

For SAMD more changes were made to keep up with the changed Pin naming
scheme, for mimxrt a clarification regarding the PWM invert argument was
added, and for rp2 a few words was spent on PWM output pairs of a
channel (aka slice).

Fix minor errors where seen.